+++ src/core/Sampler/Sampler.h
@@ -270,11 +270,12 @@ private:
 		int nComponentIdx,
 		int nBufferSize,
 		int nInitialBufferPos,
-		float cost_L,
-		float cost_R,
-		float cost_track_L,
-		float cost_track_R,
-		float fLayerPitch
+		float fGainTrack_L,
+		float fGainTrack_R,
+		float fGainJackTrack_L,
+		float fGainJackTrack_R,
+		float fLayerPitch,
+		bool bIsMuted
 	);
 
 	std::vector<std::shared_ptr<Note>> m_playingNotesQueue;