+++ CHANGELOG.md
@@ -113,6 +113,8 @@ All notable changes to this project will be documented in this file.
   - All actions can now be undone and redone.
   - Songs can be loaded by keyboard selection and pressing Enter key.
   - Playlist can be rearranged using drag and drop.
+  - Default shortcuts have been changed to be unique throughout Hydrogen and can
+    be now used in all other windows as well.
 - NSM:
   - Drumkits are no longer linked into the session folder
   - All drumkits in the session folder (proper drumkit folders containing
@@ -214,6 +216,9 @@ All notable changes to this project will be documented in this file.
   event (#1765).
 - Fix mixup of JACK per-track output ports on instrument reordering (see Changed
   section for new behaviour) (#71).
+- LADSPA aux do now respect instrument/component/layer mute and solo states.
+- Playback track does respect main volume.
+- Instrument peaks are no longer affected by main volume.
 
 ### Removed